fix: chat requests movido a chat (estaban en offer y main)
