+++ cuidaMe/urls.py
@@ -32,9 +32,6 @@ urlpatterns = [path('admin/', admin.site.urls),
     path('about/', views.about_us, name='about_us'),
     path('edit_ad/', views.edit_ad, name='edit_ad'),
     path('chat/', include('main.chat.urls', namespace='chat')),
-    path('chat/requests/', views.chat_requests_for_caregiver, name='chat_requests_for_caregiver'),
-    path('accept_chat_request/<int:chat_request_id>/', views.accept_chat_request, name='accept_chat_request'),
-    path('reject_chat_request/<int:chat_request_id>/', views.reject_chat_request, name='reject_chat_request'),
     path('pricingPlan/', views.pricing_plan, name='pricing_plan'),
     path('product_page', views.product_page, name='product_page'), #Con fines de testeo
     path('payment_successful', views.payment_successful, name='payment_successful'),
