+++ main/chat/apps.py
@@ -4,3 +4,6 @@ from django.apps import AppConfig
 class ChatConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'main.chat'
+
+    def ready(self):
+        import main.chat.signals
