+++ main/chat/templates/chat/chat_list.html
@@ -16,11 +16,11 @@
                     <p>{{ chat_request.offer.title }}</p>
                     <p>{{ chat_request.sender.location }}</p>
                         <!-- Para aceptar oferta  -->
-                    <a class="btn btn-success" href="{% url 'accept_chat_request' chat_request.id %}" role="button" style="background-color: green; color: white;">
+                    <a class="btn btn-success" href="{% url 'chat:accept_chat_request' chat_request.id %}" role="button" style="background-color: green; color: white;">
                         Aceptar
                     </a>
                     <!-- Para denegar la oferta -->
-                    <a class="btn btn-danger" href="{% url 'reject_chat_request' chat_request.id %}" role="button" style="background-color: red;color: white;">
+                    <a class="btn btn-danger" href="{% url 'chat:reject_chat_request' chat_request.id %}" role="button" style="background-color: red;color: white;">
                         Denegar
                     </a>
                     <hr style="border: 1px solid #1B4965;opacity: 0.62;position: relative;left: -20px;">
