+++ main/offer/apps.py
@@ -4,6 +4,3 @@ from django.apps import AppConfig
 class OffersConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "main.offer"
-
-    def ready(self):
-        import main.offer.signals
