+++ main/offer/models.py
@@ -2,7 +2,6 @@ from django.db import models
 from django.urls import reverse
 from django.contrib.auth.models import User
 from django.core.validators import MinValueValidator, MaxValueValidator
-from django.conf import settings
 
 class Offer(models.Model):
 
@@ -61,13 +60,3 @@ class Review(models.Model):
     def __str__(self):
         return str(self.user.username)
     
-
-class  ChatRequest(models.Model):
-    sender = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name='sent_chat_requests')
-    receiver = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, related_name='received_chat_requests')
-    offer = models.ForeignKey(Offer, on_delete=models.CASCADE)
-    accepted = models.BooleanField(default=False)
-    time_send = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return f'{self.sender} to {self.receiver}'
