+++ main/offer/templates/offers/detail.html
@@ -144,7 +144,7 @@
         </form>
     </div>
 
-                    <form method="post" action="{% url 'offer:send_chat_request' cuidador_id=offer.user.cuidador.id offer_id=offer.id %}">
+                    <form method="post" action="{% url 'chat:send_chat_request' cuidador_id=offer.user.cuidador.id offer_id=offer.id %}">
                         <div style="display: flex;align-items: center;justify-content: center;">
                             <button type="submit" style="background-color: #1B4965; border-radius: 5px; width: 373px; height: 66px;">
                                 Enviar Solicitud de Chat
