+++ main/offer/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import filterOffers, listOffers, offerDetail, rate_offer,searchOffers, publishOffer,edit_offer, send_chat_request, myOffers, delete_offer
+from .views import filterOffers, listOffers, offerDetail, rate_offer,searchOffers, publishOffer,edit_offer, myOffers, delete_offer
 
 app_name = 'offer'
 
@@ -16,7 +16,6 @@ urlpatterns = [
     path('publish/', publishOffer, name='publish'),
     path('offer/<int:offer_id>/delete/', delete_offer, name='delete_offer'),
     path('my_offers', myOffers, name='my_offers'),
-    path('send_chat_request/<int:cuidador_id>/<int:offer_id>/', send_chat_request, name='send_chat_request'),
     path('edit_offer/<int:id>', edit_offer, name='edit_offer'),
 
 ]
