+++ main/offer/views.py
@@ -7,9 +7,8 @@ from django.http import HttpResponseForbidden
 from .models import Offer, Review
 from .forms import OfferForm, ReviewForm
 import datetime
-from django.db.models import Avg
 from main.models import Cliente, Cuidador
-from .models import ChatRequest, Offer
+from .models import Offer
 from .forms import OfferForm
 import datetime
 from django.contrib import messages
@@ -149,17 +148,4 @@ def offer_detail(request, offer_id):
     offer_reviews = Review.objects.filter(offer=offer)
 
     return render(request, 'offers/detail.html', {'offer': offer, 'form': form, 'offer_reviews': offer_reviews})
-
-  
-@login_required
-def send_chat_request(request, cuidador_id, offer_id):
-    cliente = get_object_or_404(Cliente, user=request.user)
-    cuidador = get_object_or_404(Cuidador, id=cuidador_id)
-    oferta = get_object_or_404(Offer, id=offer_id)
-    # Verifica si ya existe una solicitud pendiente para esta oferta
-    existing_request = ChatRequest.objects.filter(sender=cliente.user, receiver=cuidador.user,accepted=False,offer=oferta).first()
-    if not existing_request:
-        # Si no existe, crea una nueva solicitud con la oferta asociada
-        ChatRequest.objects.create(sender=cliente.user, receiver=cuidador.user, offer=oferta)
-    return redirect('offer:list')
