+++ main/templates/main/base.html
@@ -26,6 +26,7 @@
           {% if request.user.cuidador %}
           <a class="btn" href="/offer/my_offers" role="button">My offers</a>
           <a class="btn" href="/offer/list" role="button">Offers list</a>
+          <a class="btn" href="/chat/requests" role="button">Chat requests</a>
           {% endif %}
 
           {% if request.user.cliente %}
